requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
xgboost
streamlit
matplotlib
seaborn
opencv-python-headless
shap
pillow